backend/foodgram/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Django==3.2.16
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
exceptiongroup==1.3.0
filetype==1.2.0
idna==3.10
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.11.1
python3-openid==3.2.0
pytz==2025.2
psycopg2-binary==2.9.3
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
